// src/app/model/share-performance.ts
// share performance

export class SharePerfomance {
  symbol: string;
  price: string;
  per1Month: string;
  per6Month: string;
  per1Year: string;
  per3Year: string;
  per1MonthUp: boolean = false;
  per6MonthUp: boolean = false;
  per1YearUp: boolean = false;
  per3YearUp: boolean = false;
  
  constructor() {}
}